Cucumber: move tap_row into shared_steps
